Use a real Trino server for TLS integration test
