+++ trino/etc/config.properties
@@ -5,7 +5,7 @@ http-server.http.port=8080
 discovery-server.enabled=true
 discovery.uri=http://localhost:8080
 
-http-server.authentication.type=JWT
+http-server.authentication.type=PASSWORD,JWT
 http-server.authentication.jwt.key-file=/etc/trino/secrets/public_key.pem
 http-server.https.enabled=true
 http-server.https.port=8443
