+++ trino/etc/password-authenticator.properties
@@ -1,0 +1,2 @@
+password-authenticator.name=file
+file.password-file=/etc/trino/secrets/password.db
