+++ trino/etc/secrets/password.db
@@ -1,0 +1,1 @@
+admin:$2y$10$xLVM/FUuZaNEmGOU3y7fjOLzRJOGItnTWTiFVKwIFX.ZMBctxl8gq
